Refactored java packages structure ofr Keystore Device Management Service
