+++ service/device/management/keystore/internal/src/main/java/org/eclipse/kapua/service/device/management/keystore/internal/DeviceKeystoreManagementServiceImpl.java
@@ -29,18 +29,18 @@ import org.eclipse.kapua.service.device.management.commons.AbstractDeviceManagem
 import org.eclipse.kapua.service.device.management.commons.call.DeviceCallExecutor;
 import org.eclipse.kapua.service.device.management.exception.DeviceManagementRequestContentException;
 import org.eclipse.kapua.service.device.management.keystore.DeviceKeystoreManagementService;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.request.KeystoreCertificateRequestMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.request.KeystoreCsrRequestMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.request.KeystoreKeypairRequestMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.request.KeystoreQueryRequestMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.request.KeystoreRequestChannel;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.request.KeystoreRequestPayload;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoreCsrResponseMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoreItemResponseMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoreItemsResponseMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoreNoContentResponseMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoreResponseMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoresResponseMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.request.KeystoreCertificateRequestMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.request.KeystoreCsrRequestMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.request.KeystoreKeypairRequestMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.request.KeystoreQueryRequestMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.request.KeystoreRequestChannel;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.request.KeystoreRequestPayload;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.response.KeystoreCsrResponseMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.response.KeystoreItemResponseMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.response.KeystoreItemsResponseMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.response.KeystoreNoContentResponseMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.response.KeystoreResponseMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.response.KeystoresResponseMessage;
 import org.eclipse.kapua.service.device.management.keystore.model.DeviceKeystoreCSR;
 import org.eclipse.kapua.service.device.management.keystore.model.DeviceKeystoreCSRInfo;
 import org.eclipse.kapua.service.device.management.keystore.model.DeviceKeystoreCertificate;
