+++ service/device/management/keystore/internal/src/main/java/org/eclipse/kapua/service/device/management/keystore/internal/message/request/KeystoreRequestPayload.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.device.management.keystore.message.internal.request;
+package org.eclipse.kapua.service.device.management.keystore.internal.message.request;
 
 import org.eclipse.kapua.commons.util.xml.XmlUtil;
 import org.eclipse.kapua.locator.KapuaLocator;
