+++ translator/kapua/kura/src/main/java/org/eclipse/kapua/translator/kura/kapua/keystore/TranslatorAppKeystoreCsrKuraKapua.java
@@ -14,8 +14,8 @@ package org.eclipse.kapua.translator.kura.kapua.keystore;
 
 import org.eclipse.kapua.service.device.call.kura.model.keystore.KuraKeystoreCSR;
 import org.eclipse.kapua.service.device.call.message.kura.app.response.KuraResponsePayload;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoreCsrResponseMessage;
-import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoreResponsePayload;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.response.KeystoreCsrResponseMessage;
+import org.eclipse.kapua.service.device.management.keystore.internal.message.response.KeystoreResponsePayload;
 import org.eclipse.kapua.translator.Translator;
 import org.eclipse.kapua.translator.exception.InvalidPayloadException;
 import org.slf4j.Logger;
